Была найдена проблема валидации связей сущности при удалении. Было решено содзать отдельный метод в фасаде валидатора для проверки этих связей. Данная валидация была реализована для сущности BulletinCategory как арихитектурный пример для всего остального проекта.

+++ src/BulletinBoard/Application/BulletinBoard.AppServices/Contexts/Bulletin/Services/BulletinCategoryService.cs
@@ -107,6 +107,12 @@ public sealed class BulletinCategoryService : IBulletinCategoryService
     /// <inheritdoc/>
     public async Task<bool> DeleteAsync(Guid id)
     {
+        ValidationResult validationResult = await _validator.ValidateBeforeDeletingAsync(id);
+        if (!validationResult.IsValid)
+        {
+            throw new ValidationExeption(validationResult.ToDictionary());
+        }
+
         bool isOnDeleting = await _categoryRepository.DeleteAsync(id);
 
         if (!isOnDeleting)
